Consolidate server-related types & cleaners
This just reorganizes our existing code into a common place; there should be no real changes here.

We have to leave the error types where they are for now, until we can update our Typescript generation code.
