+++ src/core/account/lobby-api.js
@@ -3,12 +3,9 @@
 import { base64 } from 'rfc4648'
 import { bridgifyObject } from 'yaob'
 
+import { type LobbyRequest } from '../../types/server-types.js'
 import { type EdgeLobby, type EdgeLoginRequest } from '../../types/types.js'
-import {
-  type LobbyRequest,
-  fetchLobbyRequest,
-  sendLobbyReply
-} from '../login/lobby.js'
+import { fetchLobbyRequest, sendLobbyReply } from '../login/lobby.js'
 import { sanitizeLoginStash, syncAccount } from '../login/login.js'
 import { getStash } from '../login/login-selectors.js'
 import { type ApiInput } from '../root-pixie.js'
