+++ src/core/login/login-stash.js
@@ -12,13 +12,16 @@ import {
 import { type Disklet, justFiles } from 'disklet'
 import { base64 } from 'rfc4648'
 
+import {
+  asEdgeBox,
+  asEdgeSnrp,
+  asPendingVoucher
+} from '../../types/server-cleaners.js'
+import { type EdgeBox, type EdgeSnrp } from '../../types/server-types.js'
 import { type EdgeLog, type EdgePendingVoucher } from '../../types/types.js'
-import { type EdgeBox, asEdgeBox } from '../../util/crypto/crypto.js'
 import { base58 } from '../../util/encoding.js'
 import { type ApiInput } from '../root-pixie.js'
-import { type EdgeSnrp, asEdgeSnrp } from '../scrypt/scrypt-pixie.js'
 import { fixUsername } from './login-selectors.js'
-import { asPendingVoucher } from './login-types.js'
 
 /**
  * The login data we store on disk.
