+++ src/core/login/login-types.js
@@ -1,43 +1,11 @@
 // @flow
 
-import { type Cleaner, asDate, asObject, asOptional, asString } from 'cleaners'
-
 import {
   type EdgePendingVoucher,
   type EdgeWalletInfo,
   type EdgeWalletInfoFull
 } from '../../types/types.js'
 import { type LoginStash } from './login-stash.js'
-
-/**
- * Data sent to authenticate with the login server.
- */
-export type LoginRequest = {
-  // The request payload:
-  data?: any,
-
-  // Common fields for all login methods:
-  deviceDescription?: string,
-  otp?: string,
-  voucherId?: string,
-  voucherAuth?: string,
-
-  // Auth key login:
-  loginId?: string,
-  loginAuth?: string,
-
-  // Password login:
-  userId?: string,
-  passwordAuth?: string,
-
-  // PIN login:
-  pin2Id?: string,
-  pin2Auth?: string,
-
-  // Recovery login:
-  recovery2Id?: string,
-  recovery2Auth?: string[]
-}
 
 // Login data decrypted into memory.
 export type LoginTree = {
@@ -88,12 +56,3 @@ export type LoginKit = {
 }
 
 export type WalletInfoFullMap = { [walletId: string]: EdgeWalletInfoFull }
-
-export const asPendingVoucher: Cleaner<EdgePendingVoucher> = asObject({
-  voucherId: asString,
-  activates: asDate,
-  created: asDate,
-  ip: asString,
-  ipDescription: asString,
-  deviceDescription: asOptional(asString)
-})
